requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
